FileTestStrategy class

FileTestStrategy class

FileTestStrategy class

Details

This class implements a test strategy that runs tests matching the source file name.

See also

Super class

muttest::TestStrategy -> FileTestStrategy

Methods

Public methods

FileTestStrategy$new()

FileTestStrategy$execute()

FileTestStrategy$clone()

Method new()

Initialize the FileTestStrategy

Usage

FileTestStrategy$new(
  load_helpers = TRUE,
  load_package = c("source", "none", "installed")
)

Arguments

load_helpers

Whether to load test helpers

load_package

The package loading strategy

Method execute()

Execute the test strategy

Usage

FileTestStrategy$execute(path, mutated_file, mutated_code, env, reporter)

Arguments

path

The path to the test directory

mutated_file

The path to the file being tested

mutated_code

The mutated code

env

The environment to run the tests in

reporter

The reporter to use for test results

Returns

The test results

Method clone()

The objects of this class are cloneable with this method.

Usage

FileTestStrategy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.