TestStrategy interface

Extend this class to implement a custom test strategy.

See also

Methods

Public methods

TestStrategy$execute()

TestStrategy$clone()

Method execute()

Execute the test strategy

Usage

TestStrategy$execute(path, plan, reporter)

Arguments

path

The path to the test directory

plan

The current mutation plan. See test_plan().

reporter

The reporter to use for test results

Returns

The test result

Method clone()

The objects of this class are cloneable with this method.

Usage

TestStrategy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.